+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 4.0.18
+### Added
+- Delete empty field groups on import with force
 ### Fixed
 - Fixed user field sources not being exported