+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Delete empty field groups on import with force
 ### Fixed
 - Fixed user field sources not being exported
+- Fixed import of multiple sites in same sitegroup
 
 ### 4.0.17 - 2018-09-24
 ### Fixed